content-security-policy/to-port-from-blink/sandbox-allow-scripts-in-http-header-expected.txt

ALERT: Script executed in iframe.
ALERT: PASS: Iframe was in a unique origin
